requirements.txt
pydantic==2.7.1
pydantic-settings==2.2.1
loguru==0.7.2
fastapi==0.111.0
uvicorn[standard]==0.29.0
openai==1.25.2
pinecone-client==4.1.0
sentence-transformers==2.7.0
pandas==2.2.2
pyspark==3.5.1 